src/Infrastructure/SFC.Players.Infrastructure.Persistence/Repositories/Repository.cs
using Microsoft.EntityFrameworkCore;

using SFC.Players.Application.Interfaces.Persistence;

namespace SFC.Players.Infrastructure.Persistence.Repositories;

public class Repository<T> : IRepository<T> where T : class
{
    protected readonly PlayersDbContext _context;

    public Repository(PlayersDbContext context) { _context = context; }

    public virtual async Task<T?> GetByIdAsync(long id)
    {
        T? t = await _context.Set<T>().FindAsync(id);
        return t;
    }

    public async Task<IReadOnlyList<T>> ListAllAsync()
    {
        return await _context.Set<T>().ToListAsync();
    }

    public async Task<T> AddAsync(T entity)
    {
        await _context.Set<T>().AddAsync(entity);
        await _context.SaveChangesAsync();

        return entity;
    }

    public async Task<T[]> AddRangeAsync(params T[] entities)
    {
        await _context.Set<T>().AddRangeAsync(entities);
        await _context.SaveChangesAsync();

        return entities;
    }

    public async Task UpdateAsync(T entity)
    {
        _context.Entry(entity).State = EntityState.Modified;
        await _context.SaveChangesAsync();
    }

    public async Task DeleteAsync(T entity)
    {
        _context.Set<T>().Remove(entity);
        await _context.SaveChangesAsync();
    }
}
